Limit scope of "users" query
With only_fields we only allow the query to access the 'id' and
'username' attributes of the UserModel. We also add a custom resolver
for isAdmin.

Notice the automatic transformation of snake_case in the backend
(is_admin) to camelCase in the frontend (isAdmin). I guess the magic
comes from Graphene?