Add mutation readBook
The training suggests adding mutations for 1) creating a readBook
relationship, 2) updating it. The read_book method already provided in
models allows us to take care of both with the same call.

With this mutation, I made double-sure that the rating calculation
works. I ran the mutation for user "me" (id=2) and book "Pride and
Prejudice", rating it 10. The book, which had a single rating of 7,
changed its average_rating to 8.5.

This commit includes a database update corresponding to said mutation.