Make rating an optional argument of readBook
The training's exercise defines rating as an optional argument of
mutation readBook. This commit takes care of that.

It also improves the average_rating logic, by excluding null values,
and moves it to models.py from schema.py.

The database update adds a HasRead without rating on the same book as
before.