Add first/last offset-based pagination on "books"
If the first kwarg is passed, last is ignored.

When last is used, offset is deduced from the end of the list. I dunno
if this is standard practice, but it works for me.

Pagination is compatible with the "fiction" filter, it's calculated
after the filter is applied.

I had to write quite a bit of defensive code for "last" pagination
because Django QuerySets don't support negative indexing.